+++ addOns/commonlib/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.4.0] - 2021-06-23
 ### Added
 - An HTTP date parser/formatter.
 
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 First version.
 
+[1.4.0]: https://github.com/zaproxy/zap-extensions/releases/commonlib-v1.4.0
 [1.3.0]: https://github.com/zaproxy/zap-extensions/releases/commonlib-v1.3.0
 [1.2.0]: https://github.com/zaproxy/zap-extensions/releases/commonlib-v1.2.0
 [1.1.0]: https://github.com/zaproxy/zap-extensions/releases/commonlib-v1.1.0